+++ test/build.info
@@ -39,7 +39,7 @@ IF[{- !$disabled{tests} -}]
           destest mdc2test \
           dhtest enginetest casttest \
           bftest ssltest_old dsatest dsa_no_digest_size_test exptest rsa_test \
-          evp_fromdata_test evp_test evp_extra_test evp_fetch_prov_test \
+          evp_pkey_provided_test evp_test evp_extra_test evp_fetch_prov_test \
           igetest v3nametest v3ext \
           crltest danetest bad_dtls_test lhash_test sparse_array_test \
           conf_include_test params_api_test params_conversion_test \
@@ -210,9 +210,9 @@ IF[{- !$disabled{tests} -}]
     DEFINE[evp_extra_test]=NO_FIPS_MODULE
   ENDIF
 
-  SOURCE[evp_fromdata_test]=evp_fromdata_test.c
-  INCLUDE[evp_fromdata_test]=../include ../apps/include
-  DEPEND[evp_fromdata_test]=../libcrypto libtestutil.a
+  SOURCE[evp_pkey_provided_test]=evp_pkey_provided_test.c
+  INCLUDE[evp_pkey_provided_test]=../include ../apps/include
+  DEPEND[evp_pkey_provided_test]=../libcrypto libtestutil.a
 
   SOURCE[igetest]=igetest.c
   INCLUDE[igetest]=../include ../apps/include
